Crate pallet_schemas

Schemas Pallet

The Schemas pallet provides functionality for handling schemas.

Config

Call

Pallet

Overview

This pallet provides an on chain repository for schemas, thereby allowing participants of the network to flexibly interact and exchange messages with each other without facing the challenge of sharing, managing and validating messages as well as schemas between them.

NOTE: In this pallet we define the payload format that is used in Messages Pallet.

The Schema pallet provides functions for:

Registering a new schema.

Setting maximum schema format size by governance.

Retrieving latest registered schema id.

Retrieving schemas by their id.

Terminology

Schema: A data structure that defines the payload format of any Message that is going to stored with that schema.

Schema Format: Serialization/Deserialization details of the schema

Schema Format Type: The type of the following Serialization/Deserialization. It can be Avro, Parquet or …

Dispatchable Functions

register_schema - Registers a new schema after some initial validation.

set_max_schema_format_bytes - Sets the maximum schemas format size (Bytes) by governance.

Genesis config

The Schemas pallet depends on the GenesisConfig.

Re-exports

pub use pallet::*;

pub use weights::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_schemas