Crate pallet_schemas

Schemas Pallet

The Schemas pallet provides functionality for handling schemas.

Configuration: Config

Extrinsics: Call

Runtime API: SchemasRuntimeApi

Custom RPC API: SchemasApiServer

Event Enum: Event

Error Enum: Error

Overview

This pallet provides an on chain repository for schemas, thereby allowing participants of the network to flexibly interact and exchange messages with each other without facing the challenge of sharing, managing and validating messages as well as schemas between them.

NOTE: In this pallet we define the payload structure that is used in Messages Pallet.

The Schema pallet provides functions for:

Registering a new schema.

Setting maximum schema model size by governance.

Retrieving latest registered schema id.

Retrieving schemas by their id.

Terminology

Schema: The structure that defines how a Message is stored and structured.

Schema Model: Serialization/Deserialization details of the schema

Schema Model Type: The type of the following Serialization/Deserialization. It can be Avro, Parquet or …

Implementations

SchemaValidator: Functions for accessing and validating Schemas. This implementation is what is used in the runtime.

SchemaProvider: Allows another pallet to resolve schema information.

Genesis config

The Schemas pallet depends on the GenesisConfig.

Re-exports

pub use pallet::*;

pub use weights::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_schemas

Structs

Schema

A structure defining a Schema